Additions and fixes to the course list
Back-end work to enable the new types of information to be displayed
(like course responsible and course state). The view was brought more
in line with current designs.

Also minor fixes to the formatting of periods (viewmodel now holds a string)
and an error in the datagen that caused all courses to be 'ongoing'.