Revamped teacher course list
The courselist in teacher details has its own viewmodel that contains
more relevant data. The view has been changed to show data closer to
the designs and now only current courses. All courses with state completed
are in a separate list in the viewmodel (not currently displayed).
The courses are now also ordered chronologically.